+++ serialization/mata/src/main/java/module-info.java
@@ -32,7 +32,6 @@ open module net.automatalib.serialization.mata {
     requires net.automatalib.api;
     requires net.automatalib.common.util;
     requires net.automatalib.core;
-    requires org.slf4j;
 
     // annotations are 'provided'-scoped and do not need to be loaded at runtime
     requires static org.checkerframework.checker.qual;