+++ serialization/mata/src/main/java/net/automatalib/serialization/mata/parser/MataNFAParser.java
@@ -22,16 +22,36 @@ import java.util.function.Function;
 
 import net.automatalib.automaton.AutomatonCreator;
 import net.automatalib.automaton.fsa.MutableNFA;
+import net.automatalib.automaton.fsa.NFA;
 import net.automatalib.common.util.IOUtil;
 import net.automatalib.exception.FormatException;
 import net.automatalib.serialization.InputModelData;
 import net.automatalib.serialization.InputModelDeserializer;
 
+/**
+ * Parser for reading {@link NFA}s from the <a
+ * href="https://github.com/VeriFIT/mata/blob/devel/AUTOMATAFORMAT.md">NFA-explicit</a> format.
+ *
+ * @param <S>
+ *         state type
+ * @param <I>
+ *         input symbol type
+ * @param <A>
+ *         concrete automaton type
+ */
 public class MataNFAParser<S, I, A extends MutableNFA<S, I>> implements InputModelDeserializer<I, A> {
 
     private final AutomatonCreator<A, I> creator;
     private final Function<String, I> symbolParser;
 
+    /**
+     * Constructor.
+     *
+     * @param creator
+     *         the creator of the concrete NFA instance
+     * @param symbolParser
+     *         the parser for transforming (string-based) labels to concrete input symbols
+     */
     public MataNFAParser(AutomatonCreator<A, I> creator, Function<String, I> symbolParser) {
         this.creator = creator;
         this.symbolParser = symbolParser;
@@ -39,17 +59,45 @@ public class MataNFAParser<S, I, A extends MutableNFA<S, I>> implements InputMod
 
     @Override
     public InputModelData<I, A> readModel(InputStream is) throws IOException, FormatException {
-
         try (Reader r = IOUtil.asNonClosingUTF8Reader(is)) {
-            final ExplicitMataParser parser = new ExplicitMataParser(r);
-
-            try {
-                parser.parse();
-            } catch (ParseException ex) {
-                throw new FormatException(ex);
-            }
-
-            return parser.extract(creator, symbolParser);
+            return parse(r, creator, symbolParser);
         }
+    }
+
+    /**
+     * Reads the contents from the given input stream and de-serializes it into a model instance.
+     *
+     * @param reader
+     *         the reader to read the contents from
+     * @param creator
+     *         the creator of the concrete NFA instance
+     * @param symbolParser
+     *         the parser for transforming (string-based) labels to concrete input symbols
+     * @param <S>
+     *         state type
+     * @param <I>
+     *         input symbol type
+     * @param <A>
+     *         concrete automaton type
+     *
+     * @return the de-serialized model data
+     *
+     * @throws FormatException
+     *         if the content of the stream was not in the expected format
+     */
+    public static <S, I, A extends MutableNFA<S, I>> InputModelData<I, A> parse(Reader reader,
+                                                                                AutomatonCreator<A, I> creator,
+                                                                                Function<String, I> symbolParser)
+            throws FormatException {
+
+        final ExplicitMataParser parser = new ExplicitMataParser(reader);
+
+        try {
+            parser.parse();
+        } catch (ParseException ex) {
+            throw new FormatException(ex);
+        }
+
+        return parser.extract(creator, symbolParser);
     }
 }